Resolve one circular import
